requirements.txt
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dlib==19.24.4
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
flask_executor==1.0.0
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
setuptools==70.0.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
sqlalchemy-utils
face_recognition